Add files via upload
The NewsVerify is a Python-based machine learning project designed to identify whether a given news article or headline is real or fake. It utilizes Natural Language Processing (NLP) techniques and libraries such as scikit-learn to preprocess text data and train a classification model. The project aims to combat misinformation by providing an automated tool that analyzes news content and predicts its authenticity with high accuracy